Фикс багов

+++ Script.py
@@ -376,12 +376,16 @@ def configupdate(zvr=None, update_time=None):
                 file.write(line)
 
 def base_update(base):
-    try:
-        base.to_excel("Database/assets.xlsx")
-    except:
-        logging.error("Открыт файл базы данных")
-    update=time.ctime(os.path.getmtime("Database/assets.xlsx"))
-    configupdate(update_time=update)
-    logging.info("Создана база годового объёма ремонта")
+    if os.path.exists("assets.xlsx"):
+        try:
+            base.to_excel("Database/assets.xlsx")
+        except:
+            logging.error("Открыт файл базы данных")
+        update=time.ctime(os.path.getmtime("Database/assets.xlsx"))
+        configupdate(update_time=update)
+        logging.info("Создана база годового объёма ремонта")
+    else:
+        shutil.copyfile('template/database/assets.xlsx', 'Database/assets.xlsx')
+
